backend/requirements.txt
Flask
requests
google-cloud-translate==2.0.1
